requirements.txt
python-telegram-bot==21.4
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
